refactor: migration.goをmain.goに戻した
